Add type_completion files
